combine srli modules into single file
